+++ src/components/bluepad32/include/uni_version.h
@@ -5,7 +5,7 @@
 #ifndef UNI_VERSION_H
 #define UNI_VERSION_H
 
-#define UNI_VERSION "4.0.2"
+#define UNI_VERSION "4.0.3"
 
 extern const char* uni_version;